migrating IEventEmitter to the same file as eventified
